Pull the prediction up to the test model page
Then we can pass it via props to the live graph and grid.
Previously we performed the prediction twice.

I've also tried to remove the knowledge of where components are rendered
from them in favour of using more focussed props.

I've also updated the connection lib in preparation for setting the LED.