Drop model status for now
If we used it consistently then the UI could wait for the model load
when it knew it was coming, but I want to focus elsewhere for now.
Right now I think starting on the test model page risks a redirect to
data-samples if the model isn't loaded in time for the React render.