Add a beta stage
This will also need CF configuration to serve it from live as we've not
deployed to the live bucket.

I think we should do the same from staging too so we can easily test the
routing.

I'm not sure this is the final answer as using the review bucket is a bit
odd but it's a safe way to make this change for now. I'd like to consider
consolidating to a single bucket for apps that have a CF function.