Add a beta stage (#627)
I'm not sure this is the final answer as using the review bucket is a bit
odd but it's a safe way to make this change for now. I'd like to consider
consolidating to a single bucket for apps that have a CF function.

https://createai.microbit.org/v/beta now works

We also got this working on https://stage-createai.microbit.org/v/beta so 
we could test the approach there first. That one isn't being invalidated in
CF when the beta branch deploys. Not sure it's worth changing that.