Use apps microbit-connection for native BT support
This integration is partial but is sufficient to use a single micro:bit to
flash data collection, record data and flash a MakeCode program.

There's lots about this that's WIP but I think it's a good base for further
fixes. Note it targets the apps branch.

I plan to merge this and then get the iOS app testable in non-dev environments
(e.g. deal with signing, internal distribution).

Known issues we're going to merge with anyway:

- Not tested on Android. The underlying connection code used to work there before recent fixes, likely still does.
- Only superficial testing on web (Mac only, V1 + V2).
  - Seems OK but we lost many subtle cases from microbit-connection that will need to be investigated there.
- No or wrong flow's error and reconnection (after 2 x fail) UX.
- Poor experience reconnecting after flashing a MakeCode hex.

Non-obvious changes:
- Switches to MakeCode programs for the data collection hex
- Upgrades Playwright and Node versions
- Very basic CLAUDE.md