Replace the data collection hex with a MakeCode program
This will significantly improve the apps experience as it will allow full
flashes to be avoided completely if you already have a MakeCode hex on the
micro:bit and in any case for the data collection - MakeCode transition.

This change also means that we now support Bluetooth mode for the data
collection hex. There's some risk of confusion and we might discuss this
further but on balance it's probably better if flashing always works the same
way.

This should significantly help development speed on the app as we can flash
over this hex.

I've added just-works variants for mobile as the UI flows really expect pairing
to happen so it's better for us not to contribute toward it unexpectedly not
happening... though we can't control this completely as it's about the hex
already on the device.

This one might be worth splitting out and applying to main but applying here
for now.

To revisit: I'd also like to merge Carlos's uhex tool and automate the
splitting of the MakeCode hex here.